+++ src/Cwaff-Misc/CwaffSynergies.cs
@@ -372,8 +372,8 @@ public static class CwaffSynergies
         NewMastery<MasteryOfPhotonCannon>(MASTERY_PHOTON_CANNON, PhotonCannon.ItemName);
         // Ammo gained from barrels is tripled, and ammo can also be gained from conventional sources.
         NewMastery<MasteryOfFuelRodGun>(MASTERY_FUEL_ROD_GUN, FuelRodGun.ItemName);
-        // TBD
-        // NewMastery<MasteryOfNightlighter>(MASTERY_NIGHTLIGHTER, Nightlighter.ItemName);
+        // Light strings pierce small obstacles and attempt to extend to other enemies upon impact.
+        NewMastery<MasteryOfNightlighter>(MASTERY_NIGHTLIGHTER, Nightlighter.ItemName);
       #endregion
 
         SanityCheckAllSynergiesHaveBeenInitialized();
@@ -681,7 +681,7 @@ internal class MasteryOfChainDriver          : MasteryDummyItem {}
 internal class MasteryOfRLTSRTSGTSBTSLB      : MasteryDummyItem {}
 internal class MasteryOfPhotonCannon         : MasteryDummyItem {}
 internal class MasteryOfFuelRodGun           : MasteryDummyItem {}
-// internal class MasteryOfNightlighter         : MasteryDummyItem {}
+internal class MasteryOfNightlighter         : MasteryDummyItem {}
 
 public enum Synergy {
     // Synergies
@@ -847,5 +847,5 @@ public enum Synergy {
     MASTERY_RLTSRTSGTSBTSLB,
     MASTERY_PHOTON_CANNON,
     MASTERY_FUEL_ROD_GUN,
-    // MASTERY_NIGHTLIGHTER,
+    MASTERY_NIGHTLIGHTER,
 };